REFACTOR: project structure
