+++ README.md
@@ -1,4 +1,10 @@
 Microservicios con Gin
 ======================
 
-Ejemplos de microservicios con Go, utilizando Gin como base para el desarrollo.
+Ejemplo de CRUD utilizando Mongo como base de datos.
+
+Para poder ejecutar este proyecto es necesario tener una instancia de mongo 
+localmente en el puerto default y establecer las siguientes variables de entorno:
+
+- MONGO_URI="mongodb://username:password@localhost:27017"
+- MONGO_DATABASE=demo
