+++ main.go
@@ -18,260 +18,41 @@ package main
 
 import (
 	"context"
-	"errors"
+	"go-microservices-example/handlers"
 	"log"
-	"net/http"
 	"os"
-	"time"
 
 	"github.com/gin-gonic/gin"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"go.mongodb.org/mongo-driver/mongo/readpref"
 )
 
 var (
-	recipes    []Recipe
-	ctx        context.Context
-	err        error
-	client     *mongo.Client
-	collection *mongo.Collection
+	ctx            context.Context
+	err            error
+	client         *mongo.Client
+	collection     *mongo.Collection
+	recipesHandler *handlers.RecipesHandler
 )
 
 func init() {
-	recipes = make([]Recipe, 0)
 	ctx = context.Background()
 	client, err = mongo.Connect(ctx, options.Client().ApplyURI(os.Getenv("MONGO_URI")))
 	if err = client.Ping(context.TODO(), readpref.Primary()); err != nil {
 		log.Fatal(err)
 	}
-	collection = client.Database(os.Getenv("MONGO_DATABASE")).Collection("recipes")
 	log.Println("Connected to MongoDB")
-}
-
-// swagger:parameters recipes newRecipe
-type Recipe struct {
-	ID           primitive.ObjectID `json:"id" bson:"_id"`
-	Name         string             `json:"name" bson:"name"`
-	Tags         []string           `json:"tags" bson:"tags"`
-	Ingredients  []string           `json:"ingredients" bson:"ingredients"`
-	Instructions []string           `json:"instructions" bson:"instructions"`
-	PublishedAt  time.Time          `json:"published_at" bson:"published_at"`
-}
-
-// swagger:operation POST /recipes recipes newRecipe
-// Create a new recipe
-// ---
-// produces:
-// - application/json
-// responses:
-//     '200':
-//         description: Successful operation
-//     '400':
-//         description: Invalid input
-func NewRecipeHandler(c *gin.Context) {
-	var recipe Recipe
-
-	if err := c.ShouldBindJSON(&recipe); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-
-		return
-	}
-
-	recipe.ID = primitive.NewObjectID()
-	recipe.PublishedAt = time.Now()
-	_, err = collection.InsertOne(ctx, recipe)
-	if err != nil {
-		log.Println(err)
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": "Error while inserting a new recipe",
-		})
-
-		return
-	}
-
-	c.JSON(http.StatusOK, recipe)
-}
-
-// swagger:operation GET /recipes recipes listRecipes
-// Returns list of recipes
-// ---
-// produces:
-// - application/json
-// responses:
-//  '200':
-//   description: Successful operation
-func ListRecipesHandler(c *gin.Context) {
-	cursor, err := collection.Find(ctx, bson.M{})
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-
-		return
-	}
-	defer cursor.Close(ctx)
-
-	recipes := make([]Recipe, 0)
-	for cursor.Next(ctx) {
-		var recipe Recipe
-		cursor.Decode(&recipe)
-		recipes = append(recipes, recipe)
-	}
-
-	c.JSON(http.StatusOK, recipes)
-}
-
-// swagger:operation PUT /recipes/{id} recipes updateRecipe
-// Update an existing recipe
-// ---
-// parameters:
-// - name: id
-//   in: path
-//   description: ID of the recipe
-//   required: true
-//   type: string
-// produces:
-// - application/json
-// responses:
-//  '200':
-//   description: Successful operation
-//  '400':
-//   description: Invalid Input
-//  '404':
-//   description: Invalid recipe ID
-func UpdateRecipeHandler(c *gin.Context) {
-	var recipe Recipe
-
-	id, err := primitive.ObjectIDFromHex(c.Param("id"))
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-
-		return
-	}
-
-	if err := c.ShouldBindJSON(&recipe); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-
-		return
-	}
-
-	_, err = collection.UpdateOne(ctx,
-		bson.M{"_id": id},
-		bson.D{{"$set", bson.D{
-			{"name", recipe.Name},
-			{"instructions", recipe.Instructions},
-			{"ingredients", recipe.Ingredients},
-			{"tags", recipe.Tags},
-		}}})
-	if err != nil {
-		log.Println(err)
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"message": "Recipe has been updated",
-	})
-}
-
-// swagger:operation DELETE /recipes/{id} recipes deleteRecipe
-// Delete an existing recipe
-// ---
-// parameters:
-// - name: id
-//   in: path
-//   description: ID of the recipe
-//   required: true
-//   type: string
-// produces:
-// - application/json
-// responses:
-//  '200':
-//   description: Successful operation
-//  '404':
-//   description: Invalid recipe ID
-func DeleteRecipeHandler(c *gin.Context) {
-	id, err := primitive.ObjectIDFromHex(c.Param("id"))
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{
-			"error": err.Error(),
-		})
-
-		return
-	}
-
-	result, err := collection.DeleteOne(ctx, bson.M{"_id": id})
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{
-			"error": err.Error(),
-		})
-
-		return
-	}
-
-	if result != nil && result.DeletedCount == 0 {
-		c.JSON(http.StatusNotFound, gin.H{
-			"error": "Recipe not found",
-		})
-
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"message": "Recipe has been deleted",
-	})
-}
-
-// swagger:operation GET /recipes/{id} recipes
-// Get a recipe by ID
-// ---
-// produces:
-// - application/json
-// parameters:
-//   - name: id
-//     in: path
-//     description: recipe ID
-//     required: true
-//     type: string
-// responses:
-//     '200':
-//         description: Successful operation
-func GetRecipeByIDHandler(c *gin.Context) {
-	objectId, _ := primitive.ObjectIDFromHex(c.Param("id"))
-	cursor := collection.FindOne(ctx, bson.M{"_id": objectId})
-	var recipe Recipe
-	err := cursor.Decode(&recipe)
-	if err != nil {
-		if errors.Is(err, mongo.ErrNoDocuments) {
-			c.JSON(http.StatusInternalServerError, gin.H{"message": "Recipe not found"})
-		} else {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		}
-
-		return
-	}
-
-	c.JSON(http.StatusOK, recipe)
+	collection = client.Database(os.Getenv("MONGO_DATABASE")).Collection("recipes")
+	recipesHandler = handlers.NewRecipesHandler(ctx, collection)
 }
 
 func main() {
 	router := gin.Default()
-	router.POST("/recipes", NewRecipeHandler)
-	router.GET("/recipes", ListRecipesHandler)
-	router.GET("/recipes/:id", GetRecipeByIDHandler)
-	router.PUT("/recipes/:id", UpdateRecipeHandler)
-	router.DELETE("/recipes/:id", DeleteRecipeHandler)
+	router.POST("/recipes", recipesHandler.CreateRecipeHandler)
+	router.GET("/recipes", recipesHandler.ListRecipesHandler)
+	router.GET("/recipes/:id", recipesHandler.GetRecipeByIDHandler)
+	router.PUT("/recipes/:id", recipesHandler.UpdateRecipeHandler)
+	router.DELETE("/recipes/:id", recipesHandler.DeleteRecipeHandler)
 	router.Run()
 }
